move create_data from test_pipeline.py to utils.py
